% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pppvalue.plot.R
\name{pppvalue.plot}
\alias{pppvalue.plot}
\title{Plot Pointwise Adjusted Permutation P-values (PPPvalue)}
\usage{
pppvalue.plot(pvalues, xind = NULL, ylab = NULL, color = NULL,
  type = NULL, lty = NULL, pch = NULL, legend = TRUE, ...)
}
\arguments{
\item{pvalues}{Output of \code{pppvalue} function.}

\item{xind}{A vector of the coordinates of points.}

\item{ylab}{A title for the y axis.}

\item{color}{The color of points or lines appearing in the plot.}

\item{type}{What type of plot should be drawn. \code{"o"} is default.}

\item{lty}{The line types for lines appearing in the legend.}

\item{pch}{The plotting symbols appearing in the plot.}

\item{legend}{logical; if \code{TRUE} the legend is shown and its position currently is fixed at \code{"bottomright"}.}

\item{...}{Other arguments to be passed to \code{plot} function}
}
\value{
Figure of pointwise adjusted p-values
}
\description{
Plot the object of the function \code{pppvalue} basically.
}
\examples{
\dontrun{
set.seed(123)
library(pppvalue)
obs <- rnorm(20,mean=2, sd=2)
perms <- matrix(rnorm(20*999),20)
pvals <- pppvalue(obs, perms)
pppvalue.plot(pvals)
}
}
\references{
Xu, M., and Reiss, P. T. (2019). Distribution-Free Pointwise Adjusted P-Values for Functional Hypotheses. \url{https://arxiv.org/abs/1912.00360}.
}
\seealso{
\code{\link{pppvalue}}
}
\author{
Meng Xu \email{mxu@campus.haifa.ac.il}
}
