% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pppvalue.R
\name{pppvalue}
\alias{pppvalue}
\title{Pointwise Adjusted Permutation P-values (PPPvalue)}
\usage{
pppvalue(realvec, permat, method = "all")
}
\arguments{
\item{realvec}{A vector of observed data.}

\item{permat}{A matrix of simulated/permuted data, the \code{nrow} of which must be the same as the length of 
\code{realvec} and the \code{ncol} of which denotes the number of the simulation or permutation.}

\item{method}{A charactor option specifying three adjustments or all of these, for \code{'single-step'}, \code{'step-down'},
\code{'ERL'} and \code{'all'}. The default is \code{'all'}. See details.}
}
\value{
Estimated pointwise adjusted p-values.
}
\description{
Perform three distribution-free adjustments for pointwise p-values based on envelope/graphical tests.
}
\details{
There are three distribution-free p-value adjustments offered in this package: 
\itemize{
\item \code{'single-step'}: the single-step adjusted p-values with respect to the GET-package;
\item \code{'step-down'}: the step-down adjusted p-values with respect to the step-down 
procedure (Westfall and Young, 1993; Ge et al., 2003; Romano and Wolf, 2016);
\item \code{'ERL'}: the adjusted p-values based on envelopes of the extreme rank length (ERL).}
Please see Xu and Reiss (2019) for more details.
}
\note{
\itemize{
\item \code{'ERL'} and \code{'all'} also offer the pointwise raw p-values w.r.t ranks and the p-value of GET in the default setting;
\item The \code{rank} adopts the default tie.method;
\item These adjustments would not consider the ties situations currently.}
}
\examples{
\dontrun{
set.seed(123)
library(pppvalue)
obs <- rnorm(20,mean=2, sd=2)
perms <- matrix(rnorm(20*999),20)
pvals <- pppvalue(obs, perms)
pppvalue.plot(pvals)
}
}
\references{
Xu, M., and Reiss, P. T. (2019). Distribution-Free Pointwise Adjusted P-Values for Functional Hypotheses. \url{https://arxiv.org/abs/1912.00360}.
}
\seealso{
The plot function \code{\link{pppvalue.plot}}. Other useful resources on Global Evelope Test refer to the GET package (\code{\link[GET]{global_envelope_test}}).
}
\author{
Meng Xu \email{mxu@campus.haifa.ac.il}, Philip Reiss
}
