% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.erl.R
\name{single.erl}
\alias{single.erl}
\title{ERL-adjusted p-values}
\usage{
single.erl(realvec, permat)
}
\arguments{
\item{realvec}{A vector of observed data.}

\item{permat}{A matrix of simulated/permuted data, the \code{nrow} of which must be the same as the length of 
\code{realvec} and the \code{ncol} of which denotes the number of the simulation or permutation.}
}
\value{
Estimated ERL-adjusted p-values.
}
\description{
Estimate the ERL adjusted p-values (internal function).
}
\seealso{
\code{\link{pppvalue}}
}
\author{
Meng Xu
}
\keyword{internal}
